requirements.txt
torch==2.1.0
torchvision==0.16.0
diffusers==0.24.0
transformers==4.37.2
opencv-python==4.9.0.80
accelerate==0.27.2
xformers==0.0.25
Pillow==10.3.0
